feat: export store utils
EX-4742
