+++ packages/x-components/src/plugins/x-plugin.types.ts
@@ -12,9 +12,9 @@ import { Store } from 'vuex';
 import { ActionsTree } from '../store/actions.types';
 import { GettersTree } from '../store/getters.types';
 import { MutationsTree } from '../store/mutations.types';
-import { StoreEmitters } from '../store/store-emitters.types';
 import { AnyXStoreModule, XStoreModule } from '../store/store.types';
-import { RequestStatus } from '../store/utils/helpers/status.helpers';
+import { RequestStatus } from '../store/utils/status-store.utils';
+import { StoreEmitters } from '../store/utils/store-emitters.utils';
 import { DeepPartial, PropsWithType } from '../utils';
 import { XEvent, XEventPayload, XEventsTypes } from '../wiring/events.types';
 import { WireMetadata, Wiring } from '../wiring/wiring.types';
