+++ packages/x-components/src/store/index.ts
@@ -1,7 +1,8 @@
 export * from './actions.types';
+export * from './utils/fetch-and-save-action.utils';
 export * from './getters.types';
 export * from './mutations.types';
-export * from './store-emitters.types';
+export * from './utils/status-store.utils';
+export * from './utils/store-emitters.utils';
 export * from './store.types';
-export * from './store.utils';
 export * from './x.module';
