+++ packages/x-components/src/store/utils/store-emitters.utils.ts
@@ -1,7 +1,7 @@
 import { WatchOptions } from 'vue';
-import { Dictionary, Returns } from '../utils';
-import { XEvent, XEventPayload } from '../wiring/events.types';
-import { AnyXStoreModule } from './store.types';
+import { Dictionary, Returns } from '../../utils/types';
+import { XEvent, XEventPayload } from '../../wiring/events.types';
+import { AnyXStoreModule } from '../store.types';
 
 /**
  * Selects a part of the store state or getters (AKA "getter" inside
@@ -69,24 +69,37 @@ export type StoreEmitters<StoreModule extends AnyXStoreModule> = {
         Returns<StoreModule['getters']>
       >;
 };
-
 /**
  * Alias for any simple state selector.
  *
  * @public
  */
 export type AnySimpleStateSelector = SimpleStateSelector<any, any, any>;
-
 /**
  * Alias for any state selector.
  *
  * @public
  */
 export type AnyStateSelector = StateSelector<any, any, any>;
-
 /**
  * Alias for any store emitters.
  *
  * @public
  */
 export type AnyStoreEmitters = StoreEmitters<AnyXStoreModule>;
+
+/**
+ * Helper function for creating type-safe {@link StoreEmitters}.
+ *
+ * @param storeModule - The store module that the emitters will be associated to.
+ * @param emitters - The {@link StoreEmitters} to create.
+ * @returns A type-safe function for storeEmitters.
+ * @public
+ */
+export function createStoreEmitters<
+  Module extends AnyXStoreModule,
+  Emitters extends StoreEmitters<Module>
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars-experimental
+>(storeModule: Module, emitters: Emitters): Emitters {
+  return emitters;
+}
