+++ packages/x-components/src/x-modules/identifier-results/store/actions/fetch-and-save-identifier-results.action.ts
@@ -1,6 +1,6 @@
 import { Result } from '@empathyco/x-types';
 // eslint-disable-next-line max-len
-import { createFetchAndSaveAction } from '../../../../store/utils/helpers/fetch-and-save-action.helpers';
+import { createFetchAndSaveAction } from '../../../../store/utils/fetch-and-save-action.utils';
 import { IdentifierResultsActionsContext } from '../types';
 
 const { fetchAndSave, cancelPrevious } = createFetchAndSaveAction<
