+++ packages/x-components/src/x-modules/query-suggestions/store/module.ts
@@ -1,4 +1,4 @@
-import { setStatus } from '../../../store/utils/helpers/status.helpers';
+import { setStatus } from '../../../store/utils/status-store.utils';
 import {
   cancelFetchAndSaveSuggestions,
   fetchAndSaveSuggestions
