+++ packages/x-components/src/x-modules/recommendations/store/actions/fetch-and-save-recommendations.action.ts
@@ -1,6 +1,6 @@
 import { Result } from '@empathyco/x-types';
 // eslint-disable-next-line max-len
-import { createFetchAndSaveAction } from '../../../../store/utils/helpers/fetch-and-save-action.helpers';
+import { createFetchAndSaveAction } from '../../../../store/utils/fetch-and-save-action.utils';
 import { RecommendationsActionContext } from '../types';
 
 const { fetchAndSave, cancelPrevious } = createFetchAndSaveAction<
