+++ packages/x-components/src/x-modules/search/store/actions/fetch-and-save-search-response.action.ts
@@ -1,6 +1,6 @@
 import { SearchResponse } from '@empathyco/x-adapter';
 // eslint-disable-next-line max-len
-import { createFetchAndSaveAction } from '../../../../store/utils/helpers/fetch-and-save-action.helpers';
+import { createFetchAndSaveAction } from '../../../../store/utils/fetch-and-save-action.utils';
 import { SearchActionContext } from '../types';
 
 const { fetchAndSave, cancelPrevious } = createFetchAndSaveAction<
