+++ packages/x-components/src/x-modules/search/store/module.ts
@@ -1,5 +1,5 @@
 import { isFacetFilter } from '@empathyco/x-types';
-import { setStatus } from '../../../store/utils/helpers/status.helpers';
+import { setStatus } from '../../../store/utils/status-store.utils';
 import { groupItemsBy } from '../../../utils/array';
 import {
   cancelFetchAndSaveSearchResponse,
